refine somecode

+++ lib/communicator/def.dart
@@ -1,6 +1,6 @@
 class Position {
-    final double x;
-    final double y;
+    final int x;
+    final int y;
     final bool isDrag;
     Position(this.x, this.y, this.isDrag);
 }
